All loss functions derived from BCE isolated into one module. Added quaterneray loss. Added viz for loss funcions
